Create separate TableBody component
