+++ components/ScrollableTableWithMenu/index.js
@@ -3,62 +3,19 @@
  * ScrollableTableWithMenu
  *
  */
-import React, { Fragment } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import { Query } from 'react-apollo';
-import gql from 'graphql-tag';
 import sortBy from 'lodash/sortBy';
-import moment from 'moment';
 import { Table, Grid } from 'semantic-ui-react';
 import { isMobile, isWindowWidthLessThanOrEqualTo } from 'lib/browser';
-import ApolloError from 'components/errors/ApolloError';
 import DashSearch from 'components/admin/DashSearch';
-import MyProjectPrimaryCol from 'components/admin/Dashboard/MyProjects/MyProjectPrimaryCol';
-import TableMobileDataToggleIcon from 'components/ScrollableTableWithMenu/TableMobileDataToggleIcon';
 import TableHeader from './TableHeader';
+import TableBody from './TableBody';
 import TableItemsDisplay from './TableItemsDisplay';
 import TableMenu from './TableMenu';
 import TableActionsMenu from './TableActionsMenu';
 import TablePagination from './TablePagination';
 import './ScrollableTableWithMenu.scss';
-
-const TEAM_VIDEO_PROJECTS_QUERY = gql`
-  query VideoProjectsByTeam( $team: String!, $first: Int ) {
-    videoProjects(
-      where: {
-        team: {
-          name: $team
-        }
-      },
-      first: $first
-     ) {
-      id
-      createdAt
-      updatedAt
-      team {
-        name
-        organization
-      }
-      author {
-        firstName
-        lastName
-      }
-      projectTitle
-      visibility
-      thumbnails {
-        url
-        alt
-        caption
-        filename
-        filetype
-        dimensions {
-          width
-          height
-        }                
-      }      
-    }
-  }
-`;
 
 /* eslint-disable react/prefer-stateless-function */
 class ScrollableTableWithMenu extends React.Component {
@@ -190,31 +147,6 @@ class ScrollableTableWithMenu extends React.Component {
     } );
   };
 
-  normalizeData = videoProjects => {
-    const normalizedVideoProjects = [];
-
-    videoProjects.forEach( videoProject => {
-      const normalizedProject = Object.create( {}, {
-        id: { value: videoProject.id },
-        createdAt: { value: moment( videoProject.createdAt ).format( 'MMMM DD, YYYY' ) },
-        updatedAt: { value: moment( videoProject.updatedAt ).format( 'MMMM DD, YYYY' ) },
-        projectTitle: { value: videoProject.projectTitle },
-        author: { value: `${videoProject.author ? videoProject.author.firstName : ''} ${videoProject.author ? videoProject.author.lastName : ''}` },
-        team: { value: videoProject.team.name },
-        visibility: { value: videoProject.visibility },
-        thumbnail: {
-          value: {
-            url: videoProject.thumbnails[0].url,
-            alt: videoProject.thumbnails[0].alt
-          }
-        }
-      } );
-      normalizedVideoProjects.push( normalizedProject );
-    } );
-
-    return normalizedVideoProjects;
-  }
-
   render() {
     const {
       tableHeaders,
@@ -262,78 +194,12 @@ class ScrollableTableWithMenu extends React.Component {
                   toggleAllItemsSelection={ this.toggleAllItemsSelection }
                   displayActionsMenu={ displayActionsMenu }
                 />
-                { /* ADD CUSTOM TABLE BODY */ }
-                <Query
-                  query={ TEAM_VIDEO_PROJECTS_QUERY }
+                <TableBody
+                  selectedItems={ selectedItems }
+                  tableHeaders={ tableHeaders }
+                  toggleItemSelection={ this.toggleItemSelection }
                   variables={ { team, first: itemsPerPage } }
-                >
-                  { ( { loading, error, data: { videoProjects } } ) => {
-                    if ( loading ) {
-                      return (
-                        <Table.Body>
-                          <Table.Row>
-                            <Table.Cell>Loading....</Table.Cell>
-                          </Table.Row>
-                        </Table.Body>
-                      );
-                    }
-                    if ( error ) {
-                      return (
-                        <Table.Body>
-                          <Table.Row>
-                            <Table.Cell>
-                              <ApolloError error={ error } />
-                            </Table.Cell>
-                          </Table.Row>
-                        </Table.Body>
-                      );
-                    }
-                    if ( !videoProjects ) return null;
-
-                    const tableData = this.normalizeData( videoProjects );
-
-                    return (
-                      <Table.Body className="myProjects">
-                        { tableData.map( d => (
-                          <Table.Row
-                            key={ d.id }
-                            className={ d.isNew ? 'myProjects_newItem' : '' }
-                          >
-                            { tableHeaders.map( ( header, i ) => (
-                              <Table.Cell
-                                data-header={ header.label }
-                                key={ `${d.id}_${header.name}` }
-                                className="items_table_item"
-                              >
-                                { i === 0
-                                  ? (
-                                    // Table must include .primary_col div for fixed column
-                                    <Fragment>
-                                      <div className="primary_col">
-                                        <MyProjectPrimaryCol
-                                          d={ d }
-                                          header={ header }
-                                          selectedItems={ selectedItems }
-                                          toggleItemSelection={ this.toggleItemSelection }
-                                        />
-                                      </div>
-                                      <TableMobileDataToggleIcon />
-                                    </Fragment>
-                                  )
-                                  : (
-                                    <span>
-                                      <div className="items_table_mobileHeader">{ header.label }</div>
-                                      { d[header.name] }
-                                    </span>
-                                  ) }
-                              </Table.Cell>
-                            ) ) }
-                          </Table.Row>
-                        ) ) }
-                      </Table.Body>
-                    );
-                  } }
-                </Query>
+                />
               </Table>
             </div>
           </Grid.Column>
@@ -355,4 +221,3 @@ ScrollableTableWithMenu.propTypes = {
 };
 
 export default ScrollableTableWithMenu;
-export { TEAM_VIDEO_PROJECTS_QUERY };
